Ignoring incorrect SSL certificate at https://weightbot.com
At 2013-08-25 the SSL certificate has expired. After that this lib stopped
working and the iPhone App stopped backuping stuff. I can't do anything with
iPhone app, but I can fix my library to get the weight data that is stored at
https://weightbot.com

Here is my bug report for the owners:
https://twitter.com/bessarabov/status/372469538350055424

And here is the link to the solution I've implemented here:
http://stackoverflow.com/questions/6795030/how-to-ignore-certificate-verify-failed-error-in-perl